1.76.1 (5.3.8p2)
Build-Id: ddfdad7f-b7c7-4f09-afd3-d6879b3e5c86
Build-Sha256: 68e03fb920ee3f740b5bd6398641acb63afafd4e9fbaca7a6debfe434af3d5a2

+++ Assembly-CSharp/DesktopCubeModelingToolsController.cs
@@ -58,7 +58,7 @@ public class DesktopCubeModelingToolsController : MonoBehaviour
 
 	private void Start()
 	{
-		ActiveTool = (CubeModelingEvent)cubeModelingStateMachine.curEvent;
+		ActiveTool = (CubeModelingEvent)(int)cubeModelingStateMachine.curEvent;
 		if (defaultTool == null)
 		{
 			SetButtonTransparency(ActiveTool);
