+++ Assembly-CSharp/NotificationTeamRequirementPanel.cs
@@ -23,7 +23,7 @@ public class NotificationTeamRequirementPanel : NotificationRequirementPanel
 	{
 		checkmark.sprite = checkmarkSprite;
 		textField.text = team.ToString();
-		switch ((MVTeam)team)
+		switch ((MVTeam)(int)team)
 		{
 		case MVTeam.Red:
 			requirementImage.sprite = redNotificationIcon;
